+++ src/Microsoft.DotNet.XHarness.CLI/Commands/GetStateCommand.cs
@@ -2,9 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.DotNet.XHarness.CLI.CommandArguments;
 using Mono.Options;
 
-namespace Microsoft.DotNet.XHarness.CLI.Common
+namespace Microsoft.DotNet.XHarness.CLI.Commands
 {
     internal abstract class GetStateCommand : XHarnessCommand
     {
